proxy: use regular user + sudo rather than root for ssh access
Other parts of podman-machine assume sudo is configured, we can make use
of it for proxy configuration as well.
This removes the assumption that ssh access will be enabled for the root
user. The proxy code is the only part of podman using ssh root access to
the podman machine VM.

Signed-off-by: Christophe Fergeau <cfergeau@redhat.com>