machine: Allow to configure user-mode networking
Currently, `podman machine init` has a `--user-mode-networking` flag
which is supposed to allow the user to enable/disable user mode
networking. When unspecified, it uses a provider-specific default,
which is true except for WSL.

This commit makes this flag functional for all providers. When the flag
is unspecified, the same provider-specific default is used.
provider-specific default is used.
Whether to use user-mode networking or not is now stored in
MachineConfig.UserModeNetworking.
VMProvider::UserModeNetworkEnabled() is renamed to
VMProvider::UserModeNetworkDefault() to better reflect
its new role.
CreateOpts::UserModeNetworking is removed as its no longer useful.