More work for --user-mode-networking
podman machine does not really support non-user mode networking except
for wsl.
This PR does more work towards that, especially for qemu, but this is
not fully complete:
- the qemu network parameters need to be different so that there is a
  real network interface instead of a unix socket backed by gvproxy
- even with a real network interface, I'm not sure this will be enough
  as there will be no dhcp server to give it an IP address. This would
  be all handled if libvirt was used...

qemu might be special in that regard, on macos, nat networking is
supported out of the box with the virtualization framework.
Not sure about the status on Windows or for krunkit.