machine: Make podman socket forwarding optional
podman-machine is forwarding some podman-specific sockets from the VM to
the host, which requires ssh access to the VM and the use of gvproxy.
This is not needed by macadam, and is currently detrimental as we want
to allow to run unmodified VM images.
This commit adds a new `Capabilities` struct to the vmconfigs, which
will be used for such optional features. podman-machine will enable
socket forwarding, but macadam will be able to opt out of it, and the
podman-machine code is modified to only try to forward sockets when the
capability is enabled.