machine: Make podman socket forwarding optional
My end goal is to make ssh access to the VM optional. This is a first
step which removes the ssh connection details from gvproxy command line.
They are problematic when there is no ssh access to the VM as gvproxy
will exit if these are specified but it can't create the connection.