allow to customize tool prefix for wsl dist
every time we create a new VM using macadam on WSL it gets prepend the
string "podman". However, this could cause some error if users try to
create two machine with the same name using both podman and macadam.

This patch allows to customize the prefix, so in macadam (and any other
tool that uses podman in the backend) we can easily prevent an
overlapping in dist name by adding a different prefix.

Signed-off-by: lstocchi <lstocchi@redhat.com>