skip applyProxies
this patch skips the ApplyProxies func with non podman machine.
This is a temporary fix to allow using bootc images with macadam.
Currently macadam fails at starting a bootc image VM because it
expects user input when executing a sudo command and, moreover, the
command tries to create folder which should not be allowed in a
read-only filesystem

Signed-off-by: Luca Stocchi <lstocchi@redhat.com>