machine: Rework sockets.DialSocketWithBackoffsAndProcCheck
This function is in my opinion a bit too specialized/complicated to use.
It has many args, but half of them are for the process check callback.

This commit replaces it with an argless `check` func which could be used
for other things than PID checks. The args can be captured in the caller
of the function.

The function is renamed to DialSocketWithBackoffsAndCheckFunc.